Avanzamos

+++ db.c
@@ -20,22 +20,31 @@ void random_dir(char *dir){
 	dir[5]='\0';
 }
 
-int db_get_hash_dir(T_db *db, char *site_id, char *hash_dir, char *site_name){
+int db_get_hash_dir(T_db *db, char *site_id, char *hash_dir,
+		    char *site_name,char *error, int *db_fail,
+		    T_logs *logs){
+
 	char query[200];
 	MYSQL_RES *result;
 	MYSQL_ROW row;
 
 	sprintf(query,"select name,hash_dir from web_site s inner join web_suscription u on (s.susc_id = u.id) where s.id=%s", site_id);
-	mysql_query(db->con,query);
-	if(result = mysql_store_result(db->con)){
-		if(row = mysql_fetch_row(result)){
-			strcpy(hash_dir,row[1]);
-			strcpy(site_name,row[0]);
-			return 1;
-		} else {
-			return 0;
-		}
+	logs_write(logs,L_DEBUG,"db_get_hash_dir",query);
+	if(mysql_query(db->con,query)){
+		*db_fail=1;
+		logs_write(logs,L_ERROR,"db_get_hash_dir","DB_ERROR");
+		return 0;
+	}
+	*db_fail = 0;
+	result = mysql_store_result(db->con);
+	if(mysql_num_rows(result) > 0){
+		result = mysql_store_result(db->con);
+		row = mysql_fetch_row(result);
+		strcpy(hash_dir,row[1]);
+		strcpy(site_name,row[0]);
+		return 1;
 	} else {
+		sprintf(error,"300|\"code\":\"302\",\"info\":\"Imposible borrar. sitio no existe\"");
 		return 0;
 	}
 }
@@ -85,7 +94,7 @@ int db_find_site(T_db *db, char *name){
 	}
 }
 
-uint16_t  db_site_exist(T_db *db, char *susc_id, char *site_id, char *error, int *db_fail){
+uint16_t  db_site_exist(T_db *db, char *susc_id, char *site_id, char *error, int *db_fail, T_logs *logs){
 	/* Si el sitio existe retorna su version. Si retorna 0 es porque
  	   no existe */
 	char query[200];
@@ -96,6 +105,7 @@ uint16_t  db_site_exist(T_db *db, char *susc_id, char *site_id, char *error, int
 	mysql_query(db->con,query);
 	result = mysql_store_result(db->con);
 	if(!result){
+        	logs_write(logs,L_ERROR,"db_site_exist", "DB_ERROR");
 		*db_fail = 1;
 		return 0;
 	} else {
@@ -111,17 +121,18 @@ uint16_t  db_site_exist(T_db *db, char *susc_id, char *site_id, char *error, int
 
 /****	Funciones de carga	****/
 
-int db_load_site_index(T_db *db, T_site *site, char *error, int *db_fail){
+int db_load_site_index(T_db *db, T_site *site, char *site_id, char *error, int *db_fail, T_logs *logs){
 	/* Carga los indices de un sitio. Previamente vacia la estructura lista_s_e */
 	char query[200];
 	MYSQL_RES *result;
 	MYSQL_ROW row;
 	T_s_e *new_index;
 	
-	sprintf(query,"select id,name from web_indexes where site_id=%s\n",row[0]);
-	printf("Consulta para obtener los indices del sitio id %s: %s\n",row[0],query);
+	sprintf(query,"select id,name from web_indexes where site_id=%s\n",site_id);
+        logs_write(logs,L_DEBUG,"db_load_sites_index", query);
 	if(mysql_query(db->con,query)){
 		/* Ocurrio un error */
+        	logs_write(logs,L_ERROR,"db_load_sites_index", "DB_ERROR");
 		*db_fail = 1;
 	} else {
 		db_fail = 0;
@@ -130,7 +141,6 @@ int db_load_site_index(T_db *db, T_site *site, char *error, int *db_fail){
 	list_s_e_clean(site_get_indexes(site));
 	result = mysql_store_result(db->con);
 	while ((row = mysql_fetch_row(result))){
-		printf("Agregando index %s\n",row[1]);
 		new_index = (T_s_e*)malloc(sizeof(T_s_e));
 		s_e_init(new_index,atoi(row[0]),row[1]);
 		list_s_e_add(site_get_indexes(site),new_index);
@@ -138,17 +148,18 @@ int db_load_site_index(T_db *db, T_site *site, char *error, int *db_fail){
 	return 1;
 }
 
-int db_load_site_alias(T_db *db, T_site *site, char *error, int *db_fail){
+int db_load_site_alias(T_db *db, T_site *site, char *site_id, char *error, int *db_fail, T_logs *logs){
 	/* Carga los alias de un sitio. Previamente vacia la estructura lista_s_e */
 	char query[200];
 	MYSQL_RES *result;
 	MYSQL_ROW row;
 	T_s_e *new_alias;
 	
-	sprintf(query,"select id,alias from web_alias where site_id=%s\n",row[0]);
-	printf("Consulta para obtener los alias del sitio id %s: %s\n",row[0],query);
+	sprintf(query,"select id,alias from web_alias where site_id=%s\n",site_id);
+        logs_write(logs,L_DEBUG,"db_load_sites_alias",query);
 	if(mysql_query(db->con,query)){
 		/* Ocurrio un error */
+        	logs_write(logs,L_ERROR,"db_load_sites_alias","DB_ERROR");
 		*db_fail = 1;
 		return 0;
 	}
@@ -157,7 +168,6 @@ int db_load_site_alias(T_db *db, T_site *site, char *error, int *db_fail){
 	list_s_e_clean(site_get_alias(site));
 	result = mysql_store_result(db->con);
 	while ((row = mysql_fetch_row(result))){
-		printf("Agregando alias %s\n",row[1]);
 		new_alias = (T_s_e*)malloc(sizeof(T_s_e));
 		s_e_init(new_alias,atoi(row[0]),row[1]);
 		list_s_e_add(site_get_alias(site),new_alias);
@@ -165,16 +175,18 @@ int db_load_site_alias(T_db *db, T_site *site, char *error, int *db_fail){
 	return 1;
 }
 
-int db_load_sites(T_db *db, T_list_site *l, char *error, int *db_fail){
+int db_load_sites(T_db *db, T_list_site *l, char *error, int *db_fail, T_logs *logs){
 	char query[200];
 	MYSQL_ROW row, row_alias;
 	T_site *new_site;
 	T_s_e *new_alias;
 
 	strcpy(query,"select s.id,name,hash_dir,version,size from web_site s inner join web_suscription p on (s.susc_id = p.id)");
+	logs_write(logs,L_DEBUG,"db_load_sites", query);
 
-	if(mysql_query(db->con,query)==0){
+	if(mysql_query(db->con,query)){
 		/* Ocurrio un error */
+		logs_write(logs,L_ERROR,"db_load_sites","DB_ERROR");
 		*db_fail = 1;
 		return 0;
 	}
@@ -184,10 +196,10 @@ int db_load_sites(T_db *db, T_list_site *l, char *error, int *db_fail){
 		new_site = (T_site*)malloc(sizeof(T_site));
 		site_init(new_site,row[1],atoi(row[0]),row[2],atoi(row[3]),atoi(row[4]));
 		/* Cargamos los alias */
-		if(!db_load_site_alias(db, new_site, error, db_fail))
+		if(!db_load_site_alias(db,new_site,row[0],error,db_fail,logs))
 			return 0;
 		/* Cargamos los indices */
-		if(!db_load_site_index(db, new_site, error, db_fail))
+		if(!db_load_site_index(db,new_site,row[0],error,db_fail,logs))
 			return 0;
 		/* Alta del sitio en la lista */
 		list_site_add(l,new_site);
@@ -196,15 +208,17 @@ int db_load_sites(T_db *db, T_list_site *l, char *error, int *db_fail){
 	return 1;
 }
 
-int db_load_workers(T_db *db, T_list_worker *l, char *error, int *db_fail){
+int db_load_workers(T_db *db, T_list_worker *l, char *error, int *db_fail, T_logs *logs){
 	char query[200];
 	MYSQL_ROW row;
 	T_worker *new_worker;
 
 	strcpy(query,"select * from web_worker");
+	logs_write(logs,L_DEBUG,"db_load_workers", query);
 
-	if(mysql_query(db->con,query)==0){
+	if(mysql_query(db->con,query)){
 		/* Ocurrio un error */
+        	logs_write(logs,L_ERROR,"db_load_workers","DB_ERROR");
 		*db_fail = 1;
 		return 0;
 	}
@@ -219,15 +233,16 @@ int db_load_workers(T_db *db, T_list_worker *l, char *error, int *db_fail){
 	return 1;
 }
 
-int db_load_proxys(T_db *db, T_list_proxy *l, char *error, int *db_fail){
+int db_load_proxys(T_db *db, T_list_proxy *l, char *error, int *db_fail, T_logs *logs){
 	char query[200];
 	MYSQL_ROW row;
 	T_proxy *new_proxy;
 
 	strcpy(query,"select id,name,ipv4,status from web_proxy");
+	logs_write(logs,L_DEBUG,"db_load_proxys", query);
 
-	if(mysql_query(db->con,query)==0){
-		/* Ocurrio un error */
+	if(mysql_query(db->con,query)){
+        	logs_write(logs,L_ERROR,"db_load_proxys","DB_ERROR");
 		*db_fail = 1;
 		return 0;
 	}
@@ -244,21 +259,26 @@ int db_load_proxys(T_db *db, T_list_proxy *l, char *error, int *db_fail){
 
 /****	Funciones para suscripciones	****/
 
-int db_susc_add(T_db *db, my_ulonglong susc_id){
+int db_susc_add(T_db *db, my_ulonglong susc_id, int *db_error, T_logs *logs){
 	char query[200];
 	char hash_dir[6];
 
 	random_dir(hash_dir);
 	sprintf(query,"insert into web_suscription values (%lu,'%s')",susc_id,hash_dir);
-	if(!mysql_query(db->con,query)){
+	logs_write(logs,L_DEBUG,"db_susc_add", query);
+	if(mysql_query(db->con,query)){
+        	logs_write(logs,L_ERROR,"db_susc_add","DB_ERROR");
+		*db_error = 1;
 		return 0;
 	}
+	*db_error = 0;
 	return 1;
 }
 
 /****	Funciones para sitios	****/
 
-int db_site_add(T_db *db, T_site **newsite, char *name, unsigned int susc_id, char *dir){
+int db_site_add(T_db *db, T_site **newsite, char *name, unsigned int susc_id,
+		char *dir, char *error, int *db_fail, T_logs *logs){
 	/* Agrega un sitio a la base de datos.
  	 * Si no pudo hacerlo retorna 0 sino 1
 	 * En el parametro dir retorna el directorio
@@ -271,33 +291,41 @@ int db_site_add(T_db *db, T_site **newsite, char *name, unsigned int susc_id, ch
 
 	sprintf(query,"insert into web_site(version,name,size,susc_id) values(1,\"%s\",1,%lu)",
 		name,susc_id,dir);
+	logs_write(logs,L_DEBUG,"db_site_add", query);
 
-	mysql_query(db->con,query);
-	if ((result = mysql_store_result(db->con)) == 0 &&
-		mysql_field_count(db->con) == 0 &&
-		mysql_insert_id(db->con) != 0){
-		
+	if(mysql_query(db->con,query)){
+		if(mysql_errno(db->con) == 1062){
+			logs_write(logs,L_ERROR,"db_site_add", "DB_ERROR");
+			*db_fail =0;
+			logs_write(logs,L_ERROR,"db_site_add","Sitio con nombre repetido");
+			strcpy(error,"300|\"code\":\"301\",\"info\":\"Ya existe sitio con ese nombre\"");
+		} else {
+			*db_fail =1;
+			logs_write(logs,L_ERROR,"db_site_add","DB_ERROR");
+		}
+		return 0;
+	} else {
 		site_id = mysql_insert_id(db->con);
 
 		/*Obtenemos el directorio */
 		sprintf(query,"select hash_dir from web_suscription where id=%i",susc_id);
 		printf("query: %s\n",query);
-		mysql_query(db->con,query);
-		result = mysql_store_result(db->con);
-		if(row = mysql_fetch_row(result)){
-			(*newsite) = (T_site *)malloc(sizeof(T_site));
-			site_init(*newsite,name,site_id,row[0],1,1);
-			strcpy(dir,row[0]);
-		} else {
+		if(mysql_query(db->con,query)){
+			logs_write(logs,L_ERROR,"db_site_add", "DB_ERROR");
+			*db_fail =1;
 			return 0;
 		}
-	} else {
-		return 0;
+		result = mysql_store_result(db->con);
+		row = mysql_fetch_row(result);
+		(*newsite) = (T_site *)malloc(sizeof(T_site));
+		site_init(*newsite,name,site_id,row[0],1,1);
+		strcpy(dir,row[0]);
+		*db_fail =0;
+		return 1;
 	}
-	return 1;
 }
 
-int db_site_mod(T_db *db, T_site *site, T_dictionary *d, char *error, int *db_fail){
+int db_site_mod(T_db *db, T_site *site, T_dictionary *d, char *error, int *db_fail, T_logs *logs){
 	/* Modifica un sitio. Si no puede retorna 0. Si puede 1. Si falla
  	   la conexion a la base de datos retorna db_fail = 1. Sino 0. Si
 	   retorna 0 y db_fail 0 entonces indica en error el motivo por el
@@ -334,6 +362,7 @@ int db_site_mod(T_db *db, T_site *site, T_dictionary *d, char *error, int *db_fa
 
 	if(mysql_query(db->con,query)){
 		/* Algo fallo */
+        	logs_write(logs,L_ERROR,"db_site_mod","DB_ERROR");
 		*db_fail = 1;
 		return 0;
 	}
@@ -351,6 +380,7 @@ int db_site_mod(T_db *db, T_site *site, T_dictionary *d, char *error, int *db_fa
 		printf("DB_SITE_MOD: %s\n",query);
 		if(mysql_query(db->con,query)){
 			/* Algo fallo */
+        		logs_write(logs,L_ERROR,"db_site_mod","DB_ERROR");
 			*db_fail = 1;
 			return 0;
 		}
@@ -369,12 +399,13 @@ int db_site_mod(T_db *db, T_site *site, T_dictionary *d, char *error, int *db_fa
 		printf("DB_SITE_MOD: %s\n",query);
 		if(mysql_query(db->con,query)){
 			/* Algo fallo */
+        		logs_write(logs,L_ERROR,"db_site_mod","DB_ERROR");
 			*db_fail = 1;
 			return 0;
 		}
 	}
 	/* Ahora modificamos los indices en el sitio */
-	db_load_site_index(db,site,error,db_fail);
+	db_load_site_index(db,site,dictionary_get(d,"site_id"),error,db_fail,logs);
 	if(db_fail) return 0;
 	
 	/* Modificamos los alias */
@@ -383,6 +414,7 @@ int db_site_mod(T_db *db, T_site *site, T_dictionary *d, char *error, int *db_fa
 		printf("DB_SITE_MOD: %s\n",query);
 		if(mysql_query(db->con,query)){
 			/* Algo fallo */
+        		logs_write(logs,L_ERROR,"db_site_mod","DB_ERROR");
 			*db_fail = 1;
 			return 0;
 		}
@@ -400,12 +432,13 @@ int db_site_mod(T_db *db, T_site *site, T_dictionary *d, char *error, int *db_fa
 		printf("DB_SITE_MOD: %s\n",query);
 		if(mysql_query(db->con,query)){
 			/* Algo fallo */
+        		logs_write(logs,L_ERROR,"db_site_mod","DB_ERROR");
 			*db_fail = 1;
 			return 0;
 		}
 	}
 	/* Ahora modificamos los alias en el sitio */
-	db_load_site_alias(db,site,error,db_fail);
+	db_load_site_alias(db,site,dictionary_get(d,"site_id"),error,db_fail,logs);
 	if(db_fail) return 0;
 
 	/* Le indicamos al sitio que se actualice en los workers */
@@ -415,15 +448,16 @@ int db_site_mod(T_db *db, T_site *site, T_dictionary *d, char *error, int *db_fa
 	return 1;
 }
 
-int db_site_status(T_db *db, char *susc_id, char *site_id, char *status, char *error, int *db_fail){
+int db_site_status(T_db *db, char *susc_id, char *site_id, char *status, char *error, int *db_fail, T_logs *logs){
 	/* Modifica el estado de un sitio */
 	char query[200];
 
-	if(!db_site_exist(db,susc_id,site_id,error,db_fail))
+	if(!db_site_exist(db,susc_id,site_id,error,db_fail,logs))
 		return 0;
 	sprintf(query,"update web_site set status=%c where user_id=%s", status,site_id);
-	if(mysql_query(db->con,query) != 0){
+	if(mysql_query(db->con,query)){
 		/* algo fallo */
+        	logs_write(logs,L_ERROR,"db_site_mod","DB_ERROR");
 		*db_fail = 1;
 		return 0;
 	}
@@ -541,7 +575,7 @@ void db_site_show(T_db *db, char **data, int *data_size, char *site_id, char *su
 	strcat(*data,"}");
 }
 
-int db_site_del(T_db *db, char *site_id){
+int db_site_del(T_db *db, char *site_id, char *error, int *db_fail, T_logs *logs){
 	/* Elimina un sitio y todas sus componentes
 	   de la base de datos dado el id */
 
@@ -549,25 +583,37 @@ int db_site_del(T_db *db, char *site_id){
 
 	/* Borramos alias */
 	sprintf(query,"delete from web_alias where site_id=%s",site_id);
-	printf("db_site_del: %s\n",query);
-	mysql_query(db->con,query);
+	logs_write(logs,L_DEBUG,"db_site_del", query);
+	if(mysql_query(db->con,query)){
+		logs_write(logs,L_ERROR,"db_site_del","DB_ERROR");
+		*db_fail=1;
+		return 0;
+	}
 
 	/* Borramos indices */
 	sprintf(query,"delete from web_indexes where site_id=%s",site_id);
-	printf("db_site_del: %s\n",query);
-	mysql_query(db->con,query);
+	logs_write(logs,L_DEBUG,"db_site_del", query);
+	if(mysql_query(db->con,query)){
+		logs_write(logs,L_ERROR,"db_site_del","DB_ERROR");
+		*db_fail=1;
+		return 0;
+	}
 
 	/* Borrar entradas del ftp */
 	/* IMPLEMENTAR */
 
 	/* Borramos sitio */
 	sprintf(query,"delete from web_site where id=%s",site_id);
-	printf("db_site_del: %s\n",query);
-	mysql_query(db->con,query);
+	logs_write(logs,L_DEBUG,"db_site_del", query);
+	if(mysql_query(db->con,query)){
+		logs_write(logs,L_ERROR,"db_site_del","DB_ERROR");
+		*db_fail=1;
+		return 0;
+	}
 	return 1;
 }
 
-int db_del_all_site(T_db *db, char *susc_id){
+int db_del_all_site(T_db *db, char *susc_id, char *error, int *db_fail, T_logs *logs){
 	/* Elimina todos los sitios de una suscripcion
  	 *  dado el id de la misma */
 	char query[200];
@@ -578,7 +624,7 @@ int db_del_all_site(T_db *db, char *susc_id){
 	mysql_query(db->con,query);
 	if(result = mysql_store_result(db->con)){
 		while(row = mysql_fetch_row(result)){
-			db_site_del(db,row[0]);
+			db_site_del(db,row[0],error,db_fail,logs);
 		}
 	} else {
 		return 0;
